+++ quantforge/__init__.py
@@ -725,6 +725,7 @@ from .bessel2 import (
 )
 from .reverse_ad import Var, reverse_gradient
 from .lbfgs import lbfgs
+from .reverse_jacobian import reverse_jacobian, reverse_hessian, reverse_gradient_vector
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1429,7 +1430,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.10.0"
+__version__ = "5.11.0"
 
 __all__ = [
     "OptionType",
@@ -2423,6 +2424,9 @@ __all__ = [
     "bessel_kn",
     "Var",
     "reverse_gradient",
+    "reverse_jacobian",
+    "reverse_hessian",
+    "reverse_gradient_vector",
     "lbfgs",
     "Dual",
     "dual_derivative",